+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v 1.3.0 (?)
 Changes in this release:
 - Add information about how to show the python interpreter information in the status bar and make it so the status bar should appear when opening an inmanta file.
+- A warning that suggests running the `inmanta project install` command is now shown when the compiler fails to install modules. (#894)
 
 # v 1.2.0 (2022-12-23)
 Changes in this release: